removes immutable rules
Previously, the immutable rules were causing builds to fail in the other
app directories after using the '--fix' option.

This commit removes the immutable rules so that --fix can be used to
help cleanup the other app repos without breaking the build.

We will add these back in at a future date.